+++ android/src/main/java/com/theoplayer/media/MediaNotificationBuilder.kt
@@ -67,6 +67,22 @@ class MediaNotificationBuilder(
     )
   )
 
+  private val forwardAction = NotificationCompat.Action(
+    R.drawable.ic_fast_forward, context.getString(R.string.fast_forward),
+    MediaButtonReceiver.buildMediaButtonPendingIntent(
+      context,
+      PlaybackStateCompat.ACTION_FAST_FORWARD
+    )
+  )
+
+  private val rewindAction = NotificationCompat.Action(
+    R.drawable.ic_rewind, context.getString(R.string.rewind),
+    MediaButtonReceiver.buildMediaButtonPendingIntent(
+      context,
+      PlaybackStateCompat.ACTION_REWIND
+    )
+  )
+
   fun build(
     @PlaybackStateCompat.State playbackState: Int,
     largeIcon: Bitmap?,
@@ -130,13 +146,15 @@ class MediaNotificationBuilder(
       // on the eyes and avoid extremely bright or fluorescent colors.
       color = ContextCompat.getColor(context, R.color.app_primary_color)
 
-      // Add a play/pause button
+      // Add play/pause, rewind and fast-forward buttons.
       if (enableMediaControls) {
+        addAction(rewindAction)
         if (playbackState == PlaybackStateCompat.STATE_PAUSED) {
           addAction(playAction)
         } else if (playbackState == PlaybackStateCompat.STATE_PLAYING) {
           addAction(pauseAction)
         }
+        addAction(forwardAction)
       } else {
         // Add empty placeholder action as clearActions() does not work.
         addAction(0, null, null)
@@ -150,7 +168,13 @@ class MediaNotificationBuilder(
       style.setMediaSession(mediaSession.sessionToken)
 
       // Add up to 3 actions to be shown in the notification's standard-sized contentView.
-      style.setShowActionsInCompactView(0)
+      if (enableMediaControls) {
+        // The Rewind, Play/Pause and FastForward actions.
+        style.setShowActionsInCompactView(0,1,2)
+      } else {
+        // The placeholder action, which was added above.
+        style.setShowActionsInCompactView(0)
+      }
 
       if (enableCancelButton) {
         // In Android 5.0 (API level 21) and later you can swipe away a notification to