+++ example/src/App.tsx
@@ -54,6 +54,7 @@ const playerConfig: PlayerConfiguration = {
     mediaSessionEnabled: true,
     skipForwardInterval: 30,
     skipBackwardInterval: 10,
+    convertSkipToSeek: true,
   },
 };